move Scoretabs to own folder, creade new component for highscore table in tooltip
