+++ client/src/components/ScoreBox.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect} from 'react';
 import { MDBContainer, MDBTabContent, MDBNav, MDBNavItem, MDBNavLink } from "mdbreact";
 import { connect } from 'react-redux';
-import ScoreTabs from './ScoreTabs'
+import ScoreTabs from './ScoreTabs/ScoreTabs'
 import Pagination from "./Pagination";
 import { isInQuery } from '../helper/utils'
 import '../styles/ScoreBox.scss'
